// encoding/src/consts/video.data.type.ts
export type VideoMetadata = {
  filename: string;
  title: string;
  artist: string;
  album: string;
  track: string;
  date: string;
  synched: boolean;
  duration: { raw: string; seconds: number };
  video: {
    container: string;
    bitrate: string;
    stream: number;
    codec: string;
    resolution: { w: number; h: number };
    resolutionSquare: { w: number; h: number };
    aspect: { x: number; y: number; string: string; value: number };
    rotate: number;
    fps: number;
    pixelString: string;
    pixel: number;
  };
  audio: {
    codec: string;
    bitrate: string;
    sample_rate: number;
    stream: number;
    channels: { raw: audioEnum; value: number };
  };
};
enum audioEnum {
  stereo = 'stereo',
  mono = 'mono',
}